Add SUASploitable VMs

+++ scripts/permissions.sh
@@ -3,7 +3,17 @@
 # Note: this must be the last script to include because you will loose root rights after running this
 
 # Set ownership for user's home directory
-chown laboratory /home/laboratory/ -R
+if id "laboratory" >/dev/null 2>&1; then
+    chown laboratory /home/laboratory/ -R
+fi
+if id "vagrant" >/dev/null 2>&1; then
+    chown vagrant /home/vagrant/ -R
+fi
 
 # Do not remove this line, removing the file prevents the user from running anything with root rights
-rm /etc/sudoers.d/laboratory
+if id "laboratory" >/dev/null 2>&1; then
+    rm /etc/sudoers.d/laboratory
+fi
+if id "vagrant" >/dev/null 2>&1; then
+    rm /etc/sudoers.d/vagrant
+fi
